create_pull_request: opt-in allow_empty draft/empty-branch mode. Factor
